대문자 메소드 소문자로 바꿨구여, jsx -> js 로 바꾸면서 폴더구조 바꿧구여, axios 추가하고 서버 연결 되는것 까지 확인했어요. 쿼리 사용법은 간단하게 notion에 올릴게요
