+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/plugin-proposal-decorators": "^7.6.0",
+    "axios": "^0.19.0",
     "customize-cra": "^0.8.0",
     "mobx": "^5.14.2",
     "mobx-react": "^6.1.4",
