+++ src/components/PriceForm.jsx
@@ -1,8 +1,8 @@
-import {userInput} from "../store";
-import {inject, observer} from "mobx-react/dist/mobx-react";
-import React, {Component} from "react";
+import { userInput } from "../store";
+import { inject, observer } from "mobx-react/dist/mobx-react";
+import React, { Component } from "react";
 
-@inject('userInput')
+@inject("userInput")
 @observer
 class PriceForm extends Component {
   constructor(props) {
@@ -13,35 +13,34 @@ class PriceForm extends Component {
   }
 
   handleChange(event) {
-    const {userInput} = this.props;
+    const { userInput } = this.props;
     userInput.setPrice(event.target.value);
   }
 
   handleSubmit(event) {
-    const {userInput} = this.props;
-    let gookbabCount = parseInt(userInput.price, 10);
-
-    const converted = gookbabCount / 5000;
-    alert("고것은 " + converted + " 국밥이네요.");
+    const { userInput } = this.props;
+    userInput.price = "";
 
     event.preventDefault();
   }
 
   render() {
-    const {userInput} = this.props;
-    return (<div>
-      <form onSubmit={this.handleSubmit}>
-        <label>
-          가격:
-          <input
-            type="text"
-            value={userInput.price}
-            onChange={this.handleChange}
-          />
-        </label>
-        <input type="submit" className="btnClass" value="확인"/>
-      </form>
-    </div>)
+    const { userInput } = this.props;
+    return (
+      <div>
+        <form onSubmit={this.handleSubmit}>
+          <label>
+            가격:
+            <input
+              type="text"
+              value={userInput.price}
+              onChange={this.handleChange}
+            />
+          </label>
+          <input type="submit" className="btnClass" value="확인" />
+        </form>
+      </div>
+    );
   }
 }
